backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
pydantic>=1.10.7
pydantic[email]>=1.10.7
python-magic>=0.4.27
Pillow>=9.5.0
PyPDF2>=3.0.1
yara-python>=4.3.0
pytest>=7.3.1
httpx>=0.24.0
